# services/api-gateway/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Alert Bot API Gateway
  description: |
    Comprehensive API Gateway for the Alert Bot system that handles authentication, routing, and webhook processing.
    
    ## Features
    - User authentication and authorization
    - JWT token management
    - API key authentication
    - TradingView webhook processing
    - Admin panel functionality
    - Rate limiting and security
    
    ## Authentication
    This API supports multiple authentication methods:
    - **JWT Bearer Token**: For user authentication
    - **API Key**: For webhook and programmatic access
    - **Admin Token**: For administrative operations
    
    ## Rate Limiting
    Different endpoints have different rate limits:
    - Authentication: 10 requests per minute
    - Webhooks: 100 requests per minute
    - General API: 60 requests per minute
    - Admin API: 30 requests per minute
  version: 1.0.0
  contact:
    name: Alert Bot API Support
    email: support@alertbot.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.alertbot.com
    description: Production server
paths:
...
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Get all users
      description: Retrieve paginated list of all users (admin only)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of users per page
        - in: query
          name: search
          schema:
            type: string
          description: Search term for email, first name, or last name
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, user]
          description: Filter by user role
        - in: query
          name: subscriptionPlan
          schema:
            type: string
            enum: [free, basic, premium]
          description: Filter by subscription plan
        - in: query
          name: isActive
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/system/stats:
    get:
      tags:
        - Admin
      summary: System statistics
      description: Get comprehensive system statistics (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStats'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for webhook and programmatic access
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        telegramId:
          type: string
          nullable: true
          description: Linked Telegram user ID
        role:
          type: string
          enum: [admin, user]
          description: User's role
        subscriptionPlan:
          type: string
          enum: [free, basic, premium]
          description: User's subscription plan
        isActive:
          type: boolean
          description: Whether the user account is active
        isEmailVerified:
          type: boolean
          description: Whether the user's email is verified
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: Last login timestamp

    # Webhook Schemas
...
    UsersListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    SystemStats:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            users:
              type: object
            alerts:
              type: object
            charts:
              type: object
            system:
              type: object
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        pages:
          type: integer
          description: Total number of pages
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        requestId:
          type: string
          description: Request identifier
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...